requirements.txt
# Scotch-RAG Core Dependencies

# Core RAG System Dependencies (Matching Original RAG System)
langchain==0.3.27
langchain-openai==0.3.29
langchain-community==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
faiss-cpu==1.8.0
sentence-transformers==2.2.2
openai==1.99.9

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas>=2.2.0
numpy>=1.26.2

# FastHTML + MonsterUI (Frontend) - Pure Python web frameworks
python-fasthtml>=0.12.24
monsterui>=1.0.25

# Web Framework
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0
pydantic>=2.7.4
pydantic-settings>=2.4.0

# File Handling
aiofiles>=24.1.0
pathlib2==2.3.7

# Utilities
tqdm==4.66.1
loguru==0.7.2
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Enhanced Document Processing (Matching Original RAG System)
llama-parse>=0.5.0
tavily-python==0.3.3

# Model Comparison Dependencies (Matching Original RAG System)
langchain-anthropic>=0.2.0
langchain-google-genai>=0.1.0
langchain-pinecone>=0.1.0
PyJWT>=2.8.0
redis>=4.5.0
unstructured>=0.10.0

# Utilities (Matching Original RAG System)
numpy>=1.26.2
pandas>=2.2.0
# aiofiles already defined above

# Rich Terminal Interface (Matching Original RAG System)
rich==13.7.0

# File Monitoring (Matching Original RAG System)
watchdog==4.0.0
